break out child components
display "zero data" error message
