+++ client/web/src/search/results/components/aggregation/AggregationChartCard.module.scss
@@ -18,79 +18,14 @@
     color: var(--text-muted);
 }
 
-// Container for both errors (business-like and internal error alerts messages)
-.error-container {
-    display: flex;
-    flex-direction: column;
-    position: relative;
-    overflow: auto;
-    padding-top: 0.5rem;
-
-    // Small mode is used for side panel (the standard size is used for the full UI)
-    &--small {
-        font-size: 0.75rem;
-    }
-}
-
 // Internal error-alert UI
 .error-alert {
     margin: 0 0 0.5rem 0;
     overflow: auto;
 }
 
-.aggregation-error-container {
-    :global(.theme-dark) & rect {
-        fill: var(--gray-09);
-    }
-
-    :global(.theme-light) & rect {
-        fill: var(--gray-03);
-    }
-
-    // Simple UI mode theme
-    :global(.core-workflow-improvements-enabled) & {
-        :global(.theme-dark) & rect {
-            fill: var(--gray-10);
-        }
-
-        :global(.theme-light) & rect {
-            fill: var(--gray-02);
-            opacity: 0.6;
-        }
-    }
-}
-// Wrapper-container for custom search availability error message UI
-.error-message-layout {
-    padding: 0 0.5rem;
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-}
-
-// Custom search availability error message UI
-.error-message {
-    margin: auto;
-    max-width: 500px;
-}
-
 .loading {
     display: flex;
     align-items: center;
     justify-content: center;
-}
-
-.zero-state-background {
-    width: 100%;
-    flex-grow: 1;
-    margin: auto;
-
-    // Visually compensate bottom spacing between error chart mock
-    // UI and aggregation controls UI which are placed below this
-    // zero state block
-    padding-bottom: 1rem;
-    padding-top: 0.5rem;
 }
