+++ client/web/src/search/results/components/aggregation/ErrorContainer.module.scss
@@ -1,0 +1,38 @@
+.aggregation-error-container {
+    :global(.theme-dark) & rect {
+        fill: var(--gray-09);
+    }
+
+    :global(.theme-light) & rect {
+        fill: var(--gray-03);
+    }
+
+    // Simple UI mode theme
+    :global(.core-workflow-improvements-enabled) & {
+        :global(.theme-dark) & rect {
+            fill: var(--gray-10);
+        }
+
+        :global(.theme-light) & rect {
+            fill: var(--gray-02);
+            opacity: 0.6;
+        }
+    }
+}
+// Wrapper-container for custom search availability error message UI
+.error-message-layout {
+    padding: 0 0.5rem;
+    position: absolute;
+    top: 0;
+    left: 0;
+    height: 100%;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+}
+
+// Custom search availability error message UI
+.error-message {
+    margin: auto;
+    max-width: 500px;
+}
